Remove automatic change to ABORT_OXIDIZER_PRESSURE (#23)
* Remove automatic change to ABORT_OXIDIZER_PRESSURE

Only allow Ground Systems to control when we go to ABORT (other than
communication errors). This was decided by everyone else because they
want control over the flight phase even when the pressure sensor is
malfunctioning. If we lose communication and the pressure goes high, we
have to wait 3 mins before we go to ABORT because there is no other way
to get there.

* Incremented software version

* Fixed comment