Config flash driver and swap debug thread jumper
Re-enable the debug thread logic, where the thread is only started
if a jumper is on AUX_1. The new logic is to start the thread if
AUX_1 is pulled to ground, and so AUX_1 now has an internal pull-up
resistor enabled on the MCU. The flash chip is tested and working.