Convert Project to C++
Note, recommended to diff with previous copies.

Used wrappers in SystemDefines.h to make the process smoother, should eventually be replaced with a nice and easy to use wrapper around the function that
should be used for all new calls to HAL. Please keep in mind the changes to HAL_SPI_Transmit where there was a change to adding const to Transmit buffers.

Verify the wrapper around HAL_UART_Transmit is valid.

Recommended to re-write the codebase from here, but converting to C++ as a basis may help for referencing code.